Move the home route into default, and the default route into collect
