+++ app/collect/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import TapModal from "@/components/modals/TapModal";
+
+export default function Page() {
+  return <TapModal />;
+}
