--- app/home/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import HomeScreen from "@/components/screens/HomeScreen";
-
-export default function Home() {
-  return <HomeScreen />;
-}
